Move Kafka directly to io submodule.
